Revert "Merge pull request #1 from TheD1rkMtr/main"
This reverts commit 466c196f1c9fb96e48c2f79b16755cd877544e77, reversing
changes made to 2b22970dcc9c03f6527e7c024dd9dcc13682e2ca.
