--- FilelessPELoader/FilelessPELoader.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32106.194
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "FilelessPELoader", "FilelessPELoader.vcxproj", "{82277B35-D159-4B44-8D54-FB66EDD58D5C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{82277B35-D159-4B44-8D54-FB66EDD58D5C}.Debug|x64.ActiveCfg = Debug|x64
-		{82277B35-D159-4B44-8D54-FB66EDD58D5C}.Debug|x64.Build.0 = Debug|x64
-		{82277B35-D159-4B44-8D54-FB66EDD58D5C}.Debug|x86.ActiveCfg = Debug|Win32
-		{82277B35-D159-4B44-8D54-FB66EDD58D5C}.Debug|x86.Build.0 = Debug|Win32
-		{82277B35-D159-4B44-8D54-FB66EDD58D5C}.Release|x64.ActiveCfg = Release|x64
-		{82277B35-D159-4B44-8D54-FB66EDD58D5C}.Release|x64.Build.0 = Release|x64
-		{82277B35-D159-4B44-8D54-FB66EDD58D5C}.Release|x86.ActiveCfg = Release|Win32
-		{82277B35-D159-4B44-8D54-FB66EDD58D5C}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9161B8CD-FAF4-43AC-9884-6995AB00EFB0}
-	EndGlobalSection
-EndGlobal
