+++ build.bat
@@ -1,0 +1,6 @@
+del "./build" /s /q
+mkdir "./build"
+echo "../build.bat" > "./build/build.bat"
+cd "./build"
+cmake ..
+cd "../"
